Change config options and serve metrics on separate port
